Move wxMoveToTrash to core
