+++ src/osx/cocoa/utils_base.mm
@@ -388,27 +388,4 @@ bool wxMacInitCocoa()
     return cocoaLoaded;
 }
 
-// Move file or directory to macOS Trash using NSFileManager.
-bool wxMoveToTrash(const wxString& path)
-{
-    wxCFStringRef cfPath(path);
-    NSURL *fileURL = [NSURL fileURLWithPath:cfPath.AsNSString()];
-    if ( fileURL == nil )
-        return false;
-
-    NSError *error = nil;
-    BOOL ok = [[NSFileManager defaultManager] trashItemAtURL:fileURL
-                                            resultingItemURL:nil
-                                                       error:&error];
-    if ( !ok )
-    {
-        wxLogError(_("'%s' couldn't be moved to trash: %s"),
-                   path,
-                   error ? wxCFStringRef::AsString([error localizedDescription])
-                         : _("unknown error"));
-    }
-
-    return ok;
-}
-
 #endif // __WXDARWIN_OSX__
